pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "solarathon"
description = "Template project for Solarathon 2023"
version = "0.0.1"
dependencies = [
    "solara==1.23.0",
    "requests",
    "pydantic",
    "yfinance",
    "mplfinance",
    "pandas_ta"
]

[project.optional-dependencies]
dev = [
    "mypy",
    "requests",
    "pydantic",
    "yfinance",
    "mplfinance",
    "pandas_ta",
]

[tool.hatch.build]
include = [
  "**/*.css",
  "**/*.py",
]
